Create bootstrap file to bootstrap dependencies components
